Revert "Merge pull request #1 from landlab/master"
This reverts commit 25d7738f4ab3c496a43a6d7a40cb987bfb6286fa, reversing
changes made to fef7e077f5a2a5c6a2c1bdc1a824e7b2cd7f5ab7.

+++ landlab/components/drainage_density/cfuncs.pyx
@@ -4,7 +4,7 @@ cimport numpy as np
 cimport cython
 
 
-DTYPE = int
+DTYPE = np.int
 ctypedef np.int_t DTYPE_INT_t
 DTYPE_FLOAT = np.double
 ctypedef np.double_t DTYPE_FLOAT_t
